backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aniso8601==9.0.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
comm==0.1.2
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.6
decorator==5.1.1
elastic-transport==8.4.0
elasticsearch==8.6.2
executing==1.2.0
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-RESTful==0.3.9
fonttools==4.38.0
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.16.1
google-cloud-speech==2.18.0
google-cloud-vision==3.3.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.21.2
ipython==8.10.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.2.0
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
multidict==6.0.4
murmurhash==1.0.9
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.24.2
openai==0.26.5
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
preshed==3.0.8
prompt-toolkit==3.0.37
proto-plus==1.22.2
protobuf==4.22.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.5
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
rsa==4.9
six==1.16.0
smart-open==6.3.0
soupsieve==2.4
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.5
stack-data==0.6.2
thinc==8.1.7
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.14
wasabi==1.1.1
wcwidth==0.2.6
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.14.0
